Move integration tests under Integration
